@@ -3,8 +3,9 @@
   "version": "0.19.1",
   "description": "Javascript Style guide for 5app",
   "main": "index.js",
+  "type": "module",
   "scripts": {
-    "lint": "eslint ./ --config eslintrc.json",
+    "lint": "eslint ./",
     "spec": "mocha test/validate.js",
     "test": "npm run lint && npm run spec"
   },
@@ -29,22 +30,18 @@
   "homepage": "https://github.com/5app/eslint-config-5app#readme",
   "devDependencies": {
     "@5app/semantic-release-config": "^1.1.0",
-    "chai": "^4.3.7",
-    "eslint": "^8.34.0",
-    "mocha": "^10.2.0",
-    "semantic-release": "^21.0.0"
-  },
-  "peerDependencies": {
-    "eslint": ">=7.15.0"
+    "@eslint/js": "^9.1.1",
+    "chai": "^5.1.0",
+    "eslint": "^9.1.1",
+    "mocha": "^10.4.0",
+    "semantic-release": "^23.0.8"
   },
   "release": {
     "extends": "@5app/semantic-release-config"
   },
   "dependencies": {
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-jsdoc": "^46.0.0",
-    "eslint-plugin-mocha": "^10.1.0",
-    "eslint-plugin-unicorn": "^47.0.0",
-    "eslint-plugin-you-dont-need-lodash-underscore": "6.12.0"
+    "eslint-plugin-jsdoc": "^48.2.3",
+    "eslint-plugin-mocha": "^10.4.3",
+    "eslint-plugin-unicorn": "^52.0.0"
   }
 }
