@@ -1,1 +1,163 @@
-module.exports = require('./eslintrc.json');
+import js from '@eslint/js';
+import jsdoc from 'eslint-plugin-jsdoc';
+import mocha from 'eslint-plugin-mocha';
+import unicorn from 'eslint-plugin-unicorn';
+// import importPlugin from "eslint-plugin-import";
+import globals from 'globals';
+// import ydnlu from "eslint-plugin-you-dont-need-lodash-underscore";
+
+export default [
+	js.configs.recommended,
+	// ydnlu.configs.all,
+	{
+		languageOptions: {
+			ecmaVersion: 'latest',
+			sourceType: 'module',
+			globals: {
+				...globals.node,
+				...globals.commonjs,
+				...globals.es6,
+			}
+		},
+		plugins: {
+			jsdoc,
+			mocha,
+			unicorn,
+			// import: importPlugin
+		},
+		rules: {
+			'arrow-spacing': [2, {before: true, after: true}],
+			'arrow-parens': [2, 'as-needed'],
+			'array-bracket-newline': [2, 'consistent'],
+			'array-bracket-spacing': [2, 'never'],
+			'array-callback-return': 2,
+			'array-element-newline': [2, 'consistent'],
+			'brace-style': [2, 'stroustrup'],
+			'comma-dangle': 0,
+			'comma-spacing': [2, {before: false, after: true}],
+			'dot-notation': 2,
+			'dot-location': [2, 'property'],
+			'eol-last': 2,
+			eqeqeq: 2,
+			'func-call-spacing': 2,
+			'func-style': ['error', 'declaration'],
+			indent: [2, 'tab', {SwitchCase: 1}],
+			'key-spacing': [2, {beforeColon: false, afterColon: true}],
+			'keyword-spacing': [2, {before: true}],
+			'linebreak-style': [2, 'unix'],
+			'max-params': [2, {max: 3}],
+			'no-alert': 2,
+			'no-await-in-loop': 2,
+			'no-caller': 2,
+			'no-cond-assign': 2,
+			'no-constant-condition': 2,
+			'no-confusing-arrow': [2, {allowParens: true}],
+			'no-console': 2,
+			'no-debugger': 2,
+			'no-div-regex': 2,
+			'no-dupe-keys': 2,
+			'no-duplicate-imports': 2,
+			'no-empty': 2,
+			'no-empty-pattern': 2,
+			'no-eval': 2,
+			'no-extend-native': 2,
+			'no-extra-bind': 2,
+			'no-extra-semi': 2,
+			'no-fallthrough': 2,
+			'no-floating-decimal': 2,
+			'no-func-assign': 2,
+			'no-global-assign': 2,
+			'no-implicit-globals': 2,
+			'no-irregular-whitespace': 2,
+			'no-mixed-spaces-and-tabs': 2,
+			'no-multi-spaces': 2,
+			'no-multi-str': 2,
+			'no-multiple-empty-lines': 2,
+			'no-new-func': 2,
+			'no-new-wrappers': 2,
+			'no-octal': 2,
+			'no-octal-escape': 2,
+			'no-param-reassign': 2,
+			'no-path-concat': 2,
+			'no-proto': 2,
+			'no-redeclare': 2,
+			'no-return-assign': 2,
+			'no-return-await': 2,
+			'no-script-url': 2,
+			'no-self-assign': 2,
+			'no-self-compare': 2,
+			'no-sequences': 2,
+			'no-spaced-func': 2,
+			'no-tabs': [2, {allowIndentationTabs: true}],
+			'no-throw-literal': 2,
+			'no-trailing-spaces': 2,
+			'no-use-before-define': [2, {functions: false}],
+			'no-useless-call': 2,
+			'no-useless-rename': 2,
+			'no-useless-return': 2,
+			'no-undef': 2,
+			'no-unexpected-multiline': 2,
+			'no-unneeded-ternary': 2,
+			'no-unreachable': 2,
+			'no-unsafe-optional-chaining': 2,
+			'no-unused-vars': [2, {caughtErrors: 'all', ignoreRestSiblings: true}],
+			'no-var': 2,
+			'object-curly-spacing': [2, 'never'],
+			'object-shorthand': 2,
+			'one-var': [2, 'never'],
+			'prefer-arrow-callback': 2,
+			'prefer-rest-params': 2,
+			'prefer-const': 2,
+			'prefer-object-has-own': 2,
+			'prefer-object-spread': 2,
+			'prefer-spread': 2,
+			'prefer-template': 2,
+			'quote-props': [2, 'as-needed'],
+			quotes: [2, 'single'],
+			radix: [2, 'as-needed'],
+			semi: [2, 'always'],
+			'semi-spacing': [2, {before: false, after: true}],
+			'space-before-blocks': 2,
+			'space-before-function-paren': [2, {
+				anonymous: 'never',
+				named: 'never',
+				asyncArrow: 'always'
+			}],
+			'spaced-comment': [2, 'always'],
+			'space-infix-ops': [2, {int32Hint: false}],
+			'space-in-parens': [2, 'never'],
+			strict: 2,
+			'template-curly-spacing': 2,
+			'valid-typeof': 2,
+			yoda: 2,
+
+			// "import/first": 2,
+
+			'jsdoc/require-jsdoc': 0,
+			'jsdoc/check-indentation': 2,
+			'jsdoc/check-param-names': 2,
+			'jsdoc/check-syntax': 2,
+			'jsdoc/check-tag-names': 2,
+			'jsdoc/check-types': 2,
+			'jsdoc/require-description': 0,
+			'jsdoc/require-hyphen-before-param-description': 2,
+			'jsdoc/require-param': 2,
+			'jsdoc/require-param-description': 2,
+			'jsdoc/require-param-name': 2,
+			'jsdoc/require-param-type': 2,
+			'jsdoc/require-returns': 2,
+			'jsdoc/require-returns-check': 2,
+			'jsdoc/require-returns-description': 2,
+			'jsdoc/require-returns-type': 2,
+			'jsdoc/valid-types': 2,
+
+			'mocha/no-exclusive-tests': 2,
+
+			'unicorn/prefer-date-now': 'error',
+			'unicorn/prefer-node-protocol': 'error',
+			'unicorn/better-regex': 'error',
+			'unicorn/numeric-separators-style': 'error'
+		},
+
+	}
+];
