@@ -1,19 +1,21 @@
 /* eslint-env mocha */
-const {ESLint} = require('eslint');
-const {expect} = require('chai');
+import {ESLint} from 'eslint';
+import assert from 'node:assert/strict';
 
-it('load config in eslint to validate all rule syntax is correct', async () => {
 
-	const cli = new ESLint({
-		useEslintrc: false,
-		overrideConfigFile: 'eslintrc.json'
-	});
+describe('Validate ESLint config', () => {
+	it('load config in eslint to validate all rule syntax is correct', async () => {
 
-	const text = `
+		const cli = new ESLint({
+			overrideConfigFile: 'index.js'
+		});
+
+		const text = `
 const foo = 1;
 function bar() {}
 bar(foo);
 `;
-	const [{errorCount}] = await cli.lintText(text);
-	expect(errorCount).to.eql(0);
+		const [{errorCount}] = await cli.lintText(text);
+		assert.strictEqual(errorCount, 0);
+	});
 });
